requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
pypdf==3.17.1
httpx==0.25.2
